ComputerVision remove non-nullable scalars
Generally the objects may be unspecified, but when specified, the scalars
contained within them, such as confidence scores, are present.

While this is a breaking change w/r/t to v1.0 of the Azure-named SDK, it is
actually an un-breaking change for our customers who have been using the
hand-crafted [SDK](https://github.com/microsoft/cognitive-vision-windows), which
has been the official SDK up to now.